requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
drf-yasg>=1.21
psycopg2-binary>=2.9
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.6
python-decouple>=3.8
django-cors-headers>=4.3
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3
